Refactor CartPage by creating smaller components for UI
