+++ src/components/CartEmptyMessage.tsx
@@ -1,0 +1,5 @@
+const CartEmptyMessage = () => {
+  return <p className="text-center">Bag is empty</p>;
+};
+
+export default CartEmptyMessage;
